Move BluetoothController to bluetooth package
