+++ audioswitch/src/main/java/com/twilio/audioswitch/AudioDeviceSelector.kt
@@ -14,6 +14,7 @@ import android.os.Build
 import android.os.Handler
 import android.util.Log
 import com.twilio.audioswitch.AudioDeviceSelector.State.STOPPED
+import com.twilio.audioswitch.bluetooth.BluetoothController
 import java.util.Collections
 import kotlin.collections.ArrayList
